requirements.txt
absl-py==0.15.0
astunparse==1.6.3
atari-py==0.2.5
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cloudpickle==1.2.2
cmake==3.22.2
cycler==0.11.0
filelock==3.6.0
flatbuffers==1.12
fonttools==4.31.2
future==0.18.2
gast==0.3.3
glcontext==2.3.5
glfw==1.12.0
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
gym==0.16.0
gym3==0.3.3
h5py==2.10.0
idna==3.3
imageio==2.16.1
imageio-ffmpeg==0.3.0
importlib-metadata==4.11.2
iniconfig==1.1.1
joblib==1.1.0
Keras==2.4.3
Keras-Applications==1.0.8
keras-contrib @ git+https://www.github.com/keras-team/keras-contrib.git@3fc5ef709e061416f4bc8a92ca3750c824b5d2b0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==13.0.0
llvmlite==0.38.0
Markdown==3.3.6
matplotlib==3.5.1
moderngl==5.6.4
mpi4py==3.1.3
numpy==1.21.6
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pluggy==1.0.0
procgen==0.10.7
protobuf==3.19.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyglet==1.5.0
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.15.0
sklearn==0.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tomli==2.0.1
typing-extensions==3.7.4.3
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.12.1
zipp==3.7.0